add prints

+++ robot_setup_tf/src/global_planner.cpp
@@ -130,7 +130,7 @@ class GlobalPlanner{
             ROS_INFO_STREAM("Map width [cells]: " << mapWidth);
             ROS_INFO_STREAM("Map height [cells]: " << mapHeight);
             ROS_INFO_STREAM("Map data size: " << mapData.size());
-            ROS_INFO_STREAM("Map data first value: " << mapData[0]);
+            ROS_INFO_STREAM("Map data first value: " << static_cast<int>(mapData[0]));
             ROS_INFO_STREAM("mapMsg->data first value: " << mapMsg->data[0]);
             
             uint32_t seed_val = 100;
